Extract TZ helpers from session.py into dsgrid/ibis/tz.py
Moves get_current_time_zone, set_current_time_zone, custom_time_zone
and set_session_time_zone out of session.py and into a dedicated
module. session.py shrinks by 53 lines and the TZ semantic (which is
already asymmetric across DuckDB / Spark) gets its own home for the
Phase 6 reconciliation work.

The new module documents the cross-backend asymmetry: on Spark,
set_current_time_zone changes how column extractions like .year() /
.hour() interpret UTC instants; on DuckDB the connection TZ only
affects TIMESTAMP WITH TIME ZONE rendering. It also notes that
custom_time_zone and set_session_time_zone are currently identical
wrappers (both names persist because callers pre-date the rename) and
should be consolidated in Phase 6.

get_runtime_session and _DuckDBRuntimeSession are imported lazily
inside the functions to avoid the session<->tz circular dependency.

Five call sites updated: annual_time_dimension_config, query_submitter,
functions (re-export hub), conftest, test_ibis_session, test_file_schema.
