+++ dsgrid/query/query_submitter.py
@@ -41,13 +41,13 @@ from dsgrid.ibis.operations import drop_columns, join, pivot
 
 from dsgrid.project import Project
 from dsgrid.ibis.session import (
-    custom_time_zone,
     read_dataframe,
     try_read_dataframe,
     write_dataframe,
     write_dataframe_and_auto_partition,
     persist_table,
 )
+from dsgrid.ibis.tz import custom_time_zone
 from dsgrid.utils.timing import timer_stats_collector, track_timing
 from dsgrid.utils.files import delete_if_exists, compute_hash, load_data
 from dsgrid.query.models import (
